+++ projects/igniteui-angular/src/lib/drop-down/drop-down-item.base.ts
@@ -1,6 +1,7 @@
 import { IDropDownBase, IGX_DROPDOWN_BASE } from './drop-down.common';
 import { Input, HostBinding, HostListener, ElementRef, Optional, Inject, DoCheck } from '@angular/core';
 import { IgxSelectionAPIService } from '../core/selection';
+import { DeprecateProperty } from '../core/deprecateDecorators';
 
 /**
  * An abstract class defining a drop-down item:
@@ -67,15 +68,15 @@ export abstract class IgxDropDownItemBase implements DoCheck {
      *
      * ```typescript
      *  let mySelectedItem = this.dropdown.selectedItem;
-     *  let isMyItemSelected = mySelectedItem.isSelected; // true
+     *  let isMyItemSelected = mySelectedItem.selected; // true
      * ```
      */
     @Input()
-    get isSelected(): boolean {
+    get selected(): boolean {
         return this._isSelected;
     }
 
-    set isSelected(value: boolean) {
+    set selected(value: boolean) {
         if (this.isHeader) {
             return;
         }
@@ -85,10 +86,27 @@ export abstract class IgxDropDownItemBase implements DoCheck {
     /**
      * @hidden
      */
+    @Input()
+    @DeprecateProperty(`IgxDropDownItemBase \`isSelected\` property is depracated.\n` +
+        `Use \`selected\` instead.`)
+    get isSelected(): boolean {
+        return this.selected;
+    }
+
+    /**
+     * @hidden
+     */
+    set isSelected(value: boolean) {
+        this.selected = value;
+    }
+
+    /**
+     * @hidden
+     */
     @HostBinding('attr.aria-selected')
     @HostBinding('class.igx-drop-down__item--selected')
     get selectedStyle(): boolean {
-        return this.isSelected;
+        return this.selected;
     }
 
     /**
@@ -99,7 +117,7 @@ export abstract class IgxDropDownItemBase implements DoCheck {
      * ```
      */
     @HostBinding('class.igx-drop-down__item--focused')
-    get isFocused(): boolean {
+    get focused(): boolean {
         return (!this.isHeader && !this.disabled) && this._isFocused;
     }
 
@@ -112,8 +130,24 @@ export abstract class IgxDropDownItemBase implements DoCheck {
      *  </igx-drop-down-item>
      * ```
      */
-    set isFocused(value: boolean) {
+    set focused(value: boolean) {
         this._isFocused = value;
+    }
+
+    /**
+     * @hidden
+     */
+    @HostBinding('class.igx-drop-down__item--focused')
+    @DeprecateProperty(`IgxDropDownItemBase \`isFocused\` property is depracated.\n` +
+        `Use \`focused\` instead.`)
+    get isFocused(): boolean {
+        return this.focused;
+    }
+    /**
+     * @hidden
+     */
+    set isFocused(value: boolean) {
+        this.focused = value;
     }
 
     /**
@@ -196,7 +230,7 @@ export abstract class IgxDropDownItemBase implements DoCheck {
     }
 
     ngDoCheck(): void {
-        if (this.isSelected) {
+        if (this.selected) {
             const dropDownSelectedItem = this.selection.first_item(this.dropDown.id);
             if (!dropDownSelectedItem || this !== dropDownSelectedItem) {
                 this.dropDown.selectItem(this);