+++ CHANGELOG.md
@@ -4,6 +4,9 @@ All notable changes for each version of this project will be documented in this 
 ## 7.2.0
 - `igxCombo`
     - **Breaking Change** `combo.value` is now only a getter.
+- `igxDropDown`
+    - `IgxDropDownItemBase` and it's descendants (of which `IgxDropDownItem`) have had their `isSelected` and `isFocused` properties **deprecated**. Instead, use `selected` and `focused` properties.
+
 ## 7.1.2
 ### Features
 - `igx-circular-bar` and `igx-linear-bar` now feature an indeterminate input property. When this property is set to true the indicator will be continually growing and shrinking along the track.